pyproject.toml
[project]
name = "isz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.2",
    "kaleido>=1.1.0",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "py-pde>=0.47.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
]
